requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
mlflow==2.22.1
google-cloud-storage
setuptools
python-dotenv==1.0.1